csv.py: move code to icubam.db
